requirements.txt
fastapi[standart]~=0.115.14
uvicorn~=0.34.3
python-jose~=3.4.0
SQLAlchemy[asyncio]~=2.0.41
pydantic~=2.11.7
asyncpg~=0.30.0
python-decouple~=3.8